requirements.txt
# FarmWisely — Python dependencies

flask==3.1.0
flask-cors==5.0.0

scikit-learn==1.6.1
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2

pymongo==4.10.1

python-dotenv==1.0.1

scipy==1.13.0

gunicorn==21.2.0